Give containers static IP addresses
This allows containers to easily talk to each other.

Docker doesn't have a command line option for specifying the IP address
of a container, but a couple "route add..." and "ip addr add..."
commands can be specified in-line as a workaround.

Right now, the client and server both have a route added so they can
communicate directly, but once the routers are added, routes will need
to be configured to actually use the routers.

See addresses.go for the static IP's chosen for the hosts in our
network.